+++ Trarizon.Library/Collections/Extensions/SpanQueries/Helpers/.cs
@@ -1,4 +1,0 @@
-﻿namespace Trarizon.Library.Collections.Extensions;
-public static partial class SpanQuery
-{
-}