+++ Trarizon.Test.Run/Program.cs
@@ -1,6 +1,7 @@
 ﻿// See https://aka.ms/new-console-template for more information
 using BenchmarkDotNet.Running;
 using System.Collections;
+using System.Diagnostics.CodeAnalysis;
 using System.Reflection;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
@@ -14,24 +15,67 @@ using Trarizon.Library.Wrappers;
 using Trarizon.Test.Run;
 using Trarizon.Test.UnitTest;
 
-if (ArrayInts().Select(i => i.ToString()).TrySingleOrNone(out var val))
-    Console.WriteLine(val.GetHashCode()); // no warn
-else
-    Console.WriteLine(val.GetHashCode()); // warn
 
-if (ArrayInts().Select(i => i.ToString()).TrySingleOrNone(out var val2, ""))
-    Console.WriteLine(val2.GetHashCode()); // no warn
-else
-    Console.WriteLine(val2.GetHashCode()); // no warn
+IFilter<int> filter = default!;
 
-if (ArrayInts().Select<int, string?>(i => i.ToString()).TrySingleOrNone(out var val3, ""))
-    Console.WriteLine(val3.GetHashCode()); // warn
-else
-    Console.WriteLine(val3.GetHashCode()); // warn
+filter.Predicate([
+    default!,
+    default!,
+])
+    .Predicate(default(Func<int, FilterResult>)!);
 
-var rest = ArrayInts().Select(i => i.ToString()).PopFirst(out var first);
-Console.WriteLine(first.GetHashCode()); // warn
-var rest2 = ArrayInts().Select(i => i.ToString()).PopFirst(out var first2, "");
-Console.WriteLine(first.GetHashCode()); // no warn
-var rest3 = ArrayInts().Select<int, string?>(i => i.ToString()).PopFirst(out var first3);
-Console.WriteLine(first.GetHashCode()); // warn
+List<int> list;
+
+
+static class FilterCreate
+{
+    public static FilterResult Success(string? err = default) => new() { Success = true, };
+
+    public static FilterResult Failed(string error) => new() { Success = false, Errors = [error], };
+}
+
+readonly struct FilterResult
+{
+    public bool Success { get; init; }
+    public string[] Errors { get; init; }
+}
+
+struct IFilter<T>
+{
+    Optional<T> value;
+    public readonly T Value => value.Value;
+
+    public readonly bool Success => value.HasValue;
+
+    public IEnumerable<string> Errors { get; private set; }
+
+    public IFilter<T> Predicate(ReadOnlySpan<Func<T, FilterResult>> filters)
+    {
+        foreach (var filter in filters) {
+            Combine(filter(Value));
+        }
+        return this;
+    }
+
+    public IFilter<T> Predicate(Func<T, FilterResult> filter)
+    {
+        Combine(filter(Value));
+        return this;
+    }
+
+    public IFilter<TResult> Select<TResult>(Func<T, TResult> selector)
+    {
+        if (Success)
+            return new() { value = selector(Value), Errors = Errors };
+        else
+            return new() { Errors = Errors };
+    }
+
+
+    private void Combine(FilterResult filter)
+    {
+        if (!filter.Success && value.HasValue)
+            value = default;
+        Errors = Errors.Concat(filter.Errors.ToArray());
+    }
+}